requirements.txt
crewai
crewai-tools
streamlit
langchain-google-genai
python-dotenv